revise project structure
