+++ app/src/main/AndroidManifest.xml
@@ -10,13 +10,13 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Projectskripsi">
         <activity
-            android:name=".ui.student.subject_detail.StSubjectActivity"
+            android:name=".module.student.subject_detail.StSubjectActivity"
             android:exported="true">
 
         </activity>
 
         <activity
-            android:name=".ui.student.main.StMainActivity"
+            android:name=".module.student.main.StMainActivity"
             android:exported="true"
             android:label="@string/title_activity_student_main">
             <intent-filter>
