+++ app/src/main/java/com/example/project_skripsi/module/student/main/home/StHomeFragment.kt
@@ -1,0 +1,42 @@
+package com.example.project_skripsi.module.student.main.home
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
+import com.example.project_skripsi.databinding.FragmentStHomeBinding
+
+class StHomeFragment : Fragment() {
+
+    private lateinit var viewModel: StHomeViewModel
+    private var _binding: FragmentStHomeBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+
+        viewModel = ViewModelProvider(this)[StHomeViewModel::class.java]
+        _binding = FragmentStHomeBinding.inflate(inflater, container, false)
+
+        viewModel.profileName.observe(viewLifecycleOwner, Observer {
+            binding.textviewProfileName.text = it
+        })
+
+        viewModel.profileClass.observe(viewLifecycleOwner, Observer {
+            binding.textviewProfileClass.text = it
+        })
+
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
