+++ app/src/main/java/com/example/project_skripsi/module/student/main/payment/StPaymentFragment.kt
@@ -1,0 +1,39 @@
+package com.example.project_skripsi.module.student.main.payment
+
+import androidx.lifecycle.ViewModelProvider
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.lifecycle.Observer
+import com.example.project_skripsi.databinding.FragmentStPaymentBinding
+
+class StPaymentFragment : Fragment() {
+
+    private lateinit var viewModel: StPaymentViewModel
+    private var _binding: FragmentStPaymentBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+
+        viewModel = ViewModelProvider(this)[StPaymentViewModel::class.java]
+        _binding = FragmentStPaymentBinding.inflate(inflater, container, false)
+
+        viewModel.text.observe(viewLifecycleOwner, Observer {
+            binding.textHome.text = it
+        })
+
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+}
