+++ app/src/main/java/com/example/project_skripsi/module/student/main/score/StScoreViewModel.kt
@@ -1,0 +1,14 @@
+package com.example.project_skripsi.module.student.main.score
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class StScoreViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is score Fragment"
+    }
+    val text: LiveData<String> = _text
+
+}
