+++ app/src/main/java/com/example/project_skripsi/module/student/subject_detail/StSubjectViewModel.kt
@@ -1,0 +1,14 @@
+package com.example.project_skripsi.module.student.subject_detail
+
+import androidx.lifecycle.ViewModel
+
+class StSubjectViewModel : ViewModel() {
+
+    companion object {
+        const val tabCount = 4
+        val tabHeader = arrayOf("Absen", "Materi", "Ujian", "Tugas")
+    }
+
+
+
+}
