+++ app/src/main/java/com/example/project_skripsi/module/student/subject_detail/resource/StSubjectResourceFragment.kt
@@ -1,0 +1,38 @@
+package com.example.project_skripsi.module.student.subject_detail.resource
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
+import com.example.project_skripsi.databinding.FragmentStSubjectResourceBinding
+
+class StSubjectResourceFragment : Fragment() {
+
+    private lateinit var viewModel: StSubjectResourceViewModel
+    private var _binding: FragmentStSubjectResourceBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+
+        viewModel = ViewModelProvider(this).get(StSubjectResourceViewModel::class.java)
+        _binding = FragmentStSubjectResourceBinding.inflate(inflater, container, false)
+
+        viewModel.text.observe(viewLifecycleOwner, Observer {
+            binding.textHome.text = it
+        })
+
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
