+++ app/src/main/res/layout/fragment_st_calendar.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.student.main.calendar.StCalendarFragment">
+    tools:context=".module.student.main.calendar.StCalendarFragment">
 
     <TextView
         android:id="@+id/text_home"
