+++ app/src/main/res/layout/fragment_st_class.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.student.main.stclass.StClassFragment">
+    tools:context=".module.student.main.stclass.StClassFragment">
 
     <TextView
         android:id="@+id/text_home"
