+++ app/src/main/res/layout/fragment_st_home.xml
@@ -51,7 +51,7 @@
 
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.student.main.home.StHomeFragment">
+        tools:context=".module.student.main.home.StHomeFragment">
 
         <LinearLayout
             android:layout_width="match_parent"
