+++ app/src/main/res/layout/fragment_st_payment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.student.main.payment.StPaymentFragment">
+    tools:context=".module.student.main.payment.StPaymentFragment">
 
     <TextView
         android:id="@+id/text_home"
