+++ app/src/main/res/layout/fragment_st_subject_resource.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.student.main.score.StScoreFragment">
+    tools:context=".module.student.main.score.StScoreFragment">
 
     <TextView
         android:id="@+id/text_home"
